+++ workflows/normalize_xml.py
@@ -12,7 +12,7 @@ thedir = os.path.dirname(__file__)
 
 # Define argument parser to pass args with script
 parser = argparse.ArgumentParser('Auto-normalize xml files')
-parser.add_argument("--indir", type=str, help='Specify the path to the directory')
+parser.add_argument("--indir", type=str, default=thedir, help='Specify the path to the directory')
 parser.add_argument("--tag", type=str, default='w', help='Specify the xml tag to process (default= "w")')
 parser.add_argument("--outdir", type=str, help='Specify the path to the saving directory')
 args = parser.parse_args()
@@ -35,8 +35,9 @@ for file in files:
     # Normalize each entry on word level
     for entry in pbar(root.findall('.//' + args.tag)).start():
         text = entry.text
-        corrected = spellux.normalize_text(text, mode="model", add_matches=True, sim_ratio=0.8, stats=False, print_unknown=False, nrule=False, indexing=False, lemmatize=False, output="string", progress=False)
-        entry.set("corr", corrected)
+        corrected = spellux.normalize_text(text, mode="model", add_matches=True, sim_ratio=0.8, stats=False, print_unknown=False, nrule=False, indexing=False, lemmatize=True, output="json", progress=False)
+        entry.set("corr", corrected[0]["correction"])
+        entry.set("lemma", corrected[0]["lemma"])
 
     name = file[4:-4] + "_normalized"
     suff = file[-4:]